requirements.txt
blinker==1.6.2
ca-certificates==2024.7.2
click==8.1.7
flask==3.0.3
freetype==2.12.1
gunicorn==22.0.0
importlib-metadata==7.0.1
itsdangerous==2.2.0
jinja2==3.1.4
jpeg==9e
lcms2==2.12
lerc==3.0
libcxx==14.0.6
libdeflate==1.17
libffi==3.4.4
libpng==1.6.39
libtiff==4.5.1
libwebp-base==1.3.2
lz4-c==1.9.4
markupsafe==2.1.3
ncurses==6.4
openjpeg==2.5.2
openssl==3.0.15
packaging==24.1
pillow==10.4.0
pip==24.2
python==3.9.19
readline==8.2
setuptools==75.1.0
sqlite==3.45.3
tk==8.6.14
tzdata==2024a
werkzeug==3.0.3
wheel==0.44.0
wtforms==2.3.3
xz==5.4.6
zipp==3.17.0
zlib==1.2.13
zstd==1.5.5
